refactor: 도메인 관련 테스트 패키지를 domain으로 이동
AutoBasedLottoGeneratorTest.java
- 테스트 패키지를 utils에서 domain으로 변경
- 동일 패키지로 인한 불필요한 import 제거

LottoNumberParserTest.java
- 테스트 패키지를 utils에서 domain으로 변경
- 중복 import 제거로 코드 정리

ManualBasedLottoGeneratorTest.java
- 테스트 패키지를 utils에서 domain으로 변경
- 불필요한 도메인 import 제거
