+++ src/test/java/lotto/domain/AutoBasedLottoGeneratorTest.java
@@ -1,9 +1,7 @@
-package lotto.utils;
+package lotto.domain;
 
 import static org.assertj.core.api.Assertions.*;
 
-import lotto.domain.AutoBasedLottoGenerator;
-import lotto.domain.LottoCount;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
