+++ src/test/java/lotto/domain/LottoNumberParserTest.java
@@ -1,11 +1,9 @@
-package lotto.utils;
+package lotto.domain;
 
 import static org.assertj.core.api.Assertions.*;
 
 import java.util.Arrays;
 import java.util.List;
-import lotto.domain.LottoNumber;
-import lotto.domain.LottoNumberParser;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
