+++ src/test/java/lotto/domain/ManualBasedLottoGeneratorTest.java
@@ -1,10 +1,8 @@
-package lotto.utils;
+package lotto.domain;
 
 import static org.assertj.core.api.Assertions.*;
 
 import java.util.List;
-import lotto.domain.Lotto;
-import lotto.domain.ManualBasedLottoGenerator;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
